requirements.txt
decord>=0.4.0
external>=0.0.1
h5py>=2.10.0
matplotlib>=3.2.2
numpy>=1.19.3
opencv_python>=4.5.1.48
PsychoPy>=2020.2.5
pypng>=0.20220715.0
scipy>=1.5.0
